+++ dsp-meta/src/api/convert/hcl/mod.rs
@@ -1,12 +1,7 @@
-mod dataset;
 mod dsp_meta_error;
-mod extracted_dataset_attributes;
-mod extracted_dataset_blocks;
-mod extracted_project_attributes;
-mod extracted_project_blocks;
+mod entity;
 pub mod hcl_attribute;
 mod hcl_block;
 pub mod hcl_body;
-mod project;
 pub(crate) mod project_metadata;
 mod value;
